refactor : 로또 구매 개수 생성에서 null 체크 추가
