+++ src/main/java/lotto/domain/lottos/LottoTicket.java
@@ -8,7 +8,7 @@ import java.util.*;
 public class LottoTicket {
 
     public static final int LOTTO_NUMBER_SIZE = 6;
-    public static final String NULL_ERROR_MESSAGE = "null값은 허용하지 않습니다.";
+    public static final String NULL_ERROR_MESSAGE = "null 값은 허용하지 않습니다.";
     public static final String EMPTY_ERROR_MESSAGE = "숫자는 하나 이상이어야 합니다.";
     public static final String COUNT_ERROR_MESSAGE = "숫자는 %d개 여야 합니다.";
     public static final String DUPLICATE_ERROR_MESSAGE = "중복된 숫자가 존재합니다.";
